requirements.txt
pandas
python-binance
python-dotenv
requests
schedule
